Create swagger json for fabric-ca REST APIs
https://jira.hyperledger.org/browse/FAB-763

This is a work in progress.
See hapi/README.md for a description.
The swagger/swagger-fabric-ca.json is the swagger doc
for the APIs.  I'm still investigating the best way to
display this doc since it doesn't convert to html well
and http://editor.swagger.io/#/ is not very good and
haven't yet found a good way to add an endpoint to
fabric-ca to display it.

Change-Id: I6c783692b6d6de9d199ed7de437de99f2312304a
Signed-off-by: Keith Smith <bksmith@us.ibm.com>