Fix fabric-ca client enroll for alpha
This fixed is needed for the enroll command to work correctly.
See fab-1918 for the command and the cryptic error which occurs.
This was also fixed in a later CR which has more than this
small change in it, so it is less risky to pull in this one
to make the enroll command work correctly in the alpha.
This works with or without the optional csr.json argument.

https://jira.hyperledger.org/browse/FAB-1918

Change-Id: Ia3ee349f29ee06c481064526f89a2253795af024
Signed-off-by: Keith Smith <bksmith@us.ibm.com>