Fix fabric-ca client enroll for alpha
This fixed is required for alpha.
It simply removes the option of specifying a csr.json
file for the client enroll command.
Without it, we would have to always use the "-port" option
on the fabric-ca client enroll command and will cause confusion
for alpha, since the error message when not doing so is confusing.

https://jira.hyperledger.org/browse/FAB-1918

Change-Id: Ia3ee349f29ee06c481064526f89a2253795af024
Signed-off-by: Keith Smith <bksmith@us.ibm.com>