fabric-ca-server start for cobra/viper CLI
This change set is a continuation of the rebase of fabric-ca CLI
on cobra/viper.  This change set implements the first part of the
   fabric-ca-server start
command.  From a configuration perspective, the server start
command is the most intrusive because the cfssl CLI permeated
the code.  The lib/server.go now is a proper library API which
is independent of cfssl or even viper config.  The cobra/viper
specific calls are intentionally all in cmd/fabric-ca-server.

See https://jira.hyperledger.org/browse/FAB-2012

Change-Id: Iac6413c2de115fd66aada121cbfdbce22b11d4e9
Signed-off-by: Keith Smith <bksmith@us.ibm.com>