Complete fabric-ca-server start command
This is a continuation of the cobra/viper work outlined in
https://jira.hyperledger.org/browse/FAB-2012

This completes the work on the "fabric-ca-server start" command
so that we can easily start the server with the 5 endpoints enabled:
register, enroll, tcert, reenroll, and revoke.

The main logic is in lib/server.go, which delegates to the endpoint handlers.
I moved all of the endpoint handlers from cli/server to lib and the remaining
code in cli/server references it there.

Everything under cli will eventually go away, some time after we have completed the
fabric-ca-server and fabric-ca-client and deemed safe to remove.

Change-Id: I6bfe3f643d8141902f33c46d4639aaebb93268c6
Signed-off-by: Keith Smith <bksmith@us.ibm.com>