fabric-ca-client commands for cobra/viper CLI
This change set is a continuation of the rebase of fabric-ca CLI
on cobra/viper.  This change set implements fabric-ca-client commands.

Default behavior changed to use http when connecting to server. Test
cases will be added for TLS through new command line in a future
change set using new server start and stop feature introduced as part
of the new fabric-ca server CLI.

Change-Id: I669ec58287447705e7cbf938c198db4d82135ca6
Signed-off-by: Saad Karim <skarim@us.ibm.com>