Intermediate CA server support (part 1)
https://jira.hyperledger.org/browse/FAB-1465

This is the 1st part of adding intermediate CA server support for
fabric-ca-server.  In particular, it implements the Enroll func
on the ClientConfig object.  This will be used in the following
change set to complete the implementation of 1465 by calling this
func from the server init command if a URL of a parent server is
provided.

Change-Id: I6ce6defbd0344c2f0bf64d0bd038154627008e0f
Signed-off-by: Keith Smith <bksmith@us.ibm.com>