Registration request part of client config
Registration process simplified by incorporate registration
information in the overall client config file. This eliminates having
to provide a register file through a flag in the register CLI.
Registrar will update this config file with any relevant user
information before calling the register request from the command line.

This will be followed up with another change-set that will force
the client to only be able to execute enroll command as it first
command. This will create the enrollment data, client config file,
and client home directory. This is all pertinent information that is
needed for other commands to complete successfully.

https://jira.hyperledger.org/browse/FAB-2233

Change-Id: I61bdabe75a7d14425864e9eed2be6dc771f8d31e
Signed-off-by: Saad Karim <skarim@us.ibm.com>