Reflect to add server config flags
http://jira.hyperledger.org/browse/FAB-2012

This change set uses reflection to add flags for the lib.ServerConfig
structure.  As we need to add new fields to the config, the flags will
be automatically added based on field tags.

The main reflection code is in util/struct.go, and the flag specific
code is RegisterFlags in util/flag.go.  The cmd/fabric-ca-server/main.go
then calls the RegisterFlags method with the lib/ServerConfig struct.

Change-Id: Ie0208b0817e9726536182c297851e35b994e195e
Signed-off-by: Keith Smith <bksmith@us.ibm.com>