Registrar can configure max enrollment for user
Registrar registering a user should be able to set max enrollment for
user. The max enrollment value that the registrar can set for a new
user will be less than or equal to the max enrollment value defined in
the Fabric-ca server configuration file.

Also, made some logging improvements for better readability and
debugging.

https://jira.hyperledger.org/browse/FAB-1545

Change-Id: I452d9d6c52eab31b72fa9394d9ca0ab160619501
Signed-off-by: Saad Karim <skarim@us.ibm.com>