Base 64 encode/decode with padding
See https://jira.hyperledger.org/browse/FAB-2491

The problem is that the standard base 64 encoding is with padding,
but fabric-ca was using the raw standard encoding which does not
include padding characters.  The change is to encode with base 64
standard WITH PADDING, and when decoding, try with padding 1st but
if it fails, try without padding.

Change-Id: I43d3456a6ade5cf3f902c0127c1380cbc0dc066d
Signed-off-by: Keith Smith <bksmith@us.ibm.com>