Remove cli from fabric-ca
This change set removes the cli directory from fabric-ca.
The cli directory contained all of the code for the old
fabric-ca executable, but now that fabric-ca-server and
fabric-ca-client executables are functional, it is time
to delete the old code.

This also required updating some test code also which was
still using the code under cli.

https://jira.hyperledger.org/browse/FAB-2012

Change-Id: I4b8dc16e2be39b9c27a52bb167e285e0b88adcf5
Signed-off-by: Keith Smith <bksmith@us.ibm.com>