Remove global variables in lib
https://jira.hyperledger.org/browse/FAB-2012

This is the last change set pertaining to the cobra/viper restructure.
It removes all of the server-related global variables which were used
in the past.  The server is now implemented as a true library whose state
variables are in the server or server config objects

The primary changes are to pass the server object to all of the
server endpoint objects (i.e. those of the form serverXXXX.go) so that
they can reference the per-server state variables.
I also had to move some of the util.go funtions into server
so they had access to these per-server variables.

Change-Id: I1208542b6aedea30bedcad280808429622900aff
Signed-off-by: Keith Smith <bksmith@us.ibm.com>