FAB-2571: FVT for split client/server executables
Updated the existing fvt tests for the new cli commands.

So that the fvt can start running in the CI, tests that are
expected to fail have been placed in staging directory.
This includes tests that
 - are currently failing with known issues
 - need additional code to function properly
Hopefully, this will make it easier to en/disable tests
in the future.

Deleted some convenience stand-alone script files and put
the functionality into the fabric-ca_utils file .

Deleted tests that have been obsoleted by the new configuration.

Updated the Makefile to re-enable the docker-fvt images.

Change-Id: I64a21d308db2be8dd2a9be93ee46a52b1c29784c
Signed-off-by: rennman <eabailey@us.ibm.com>