FAB-2571: FVT for split client/server executables
Updated the existing fvt tests for the new cli commands.

So that the fvt can start running in the CI, tests that are
expected to fail have been placed in staging directory.
This includes tests that
 - are currently failing with known issues
 - need additional code to function properly
Hopefully, this will make it easier to en/disable tests
in the future.

Deleted some convenience stand-alone script files and put
the functionality into the fabric-ca_utils file .

Deleted tests that have been obsoleted by the new configuration.

Updated the Makefile to re-enable the docker-fvt images.

Change-Id: I64a21d308db2be8dd2a9be93ee46a52b1c29784c
Signed-off-by: rennman <eabailey@us.ibm.com>

Improvements to revoke client side command

Revocation logic improved to check that revoker has
the necessary affiliation to revoke a user.

Check added to make sure that either the enrollment ID
is provided or both serial and AKI are provided, all three
can't be provided at the same time.

Storing both serial and AKI in database in hex format.
This improves user experience as openssl returns both serial
and AKI in hex format. User's can now directly take the values
and pass into the revoke command without having to do any
manipulation.

Test cases added to test various use cases for revocation

https://jira.hyperledger.org/browse/FAB-2517

Change-Id: Ice68cfcc7b134046a18d2124ec3c3b0b1ae8be68
Signed-off-by: Saad Karim <skarim@us.ibm.com>
Signed-off-by: Keith Smith <bksmith@us.ibm.com>

FAB-2571: FVT for split client/server executables

Updated the existing fvt tests for the new cli commands.

So that the fvt can start running in the CI, tests that are
expected to fail have been placed in staging directory.
This includes tests that
 - are currently failing with known issues
 - need additional code to function properly
Hopefully, this will make it easier to en/disable tests
in the future.

Deleted some convenience stand-alone script files and put the
functionality into the fabric-ca_utils file . Deleted tests that
have been obsoleted by the new configuration.

Updated the Makefile to re-enable the docker-fvt images.

Updated the scripts/run_ldap_tests to fix the re-located ldap directory.

Enabled all of the backend DB variations.

Change-Id: I337599c663cf6dacc3c9644ade76a01fc4a1ff17
Signed-off-by: Allen Bailey <eabailey@us.ibm.com>