Revoked user should not be able to make requests
Once a user has been revoked or if it's certificate has
been revoked, it should not be able to continue to make requests
such as requesting a tcert batch or registering a new user.

This will add the necessary logic to check the status of the
certificate in the token to make sure that it is not revoked.

https://jira.hyperledger.org/browse/FAB-2668

Change-Id: I6b9d66796d968cf44e2aeb49fd3324ea2e75a7e8
Signed-off-by: Saad Karim <skarim@us.ibm.com>