Add support for -M option for enroll/reenroll
See https://jira.hyperledger.org/browse/FAB-2602

This change set is the 2nd and last part of FAB-2602.
In particular, it adds support for the "-M <mspDir>"
option for the "fabric-ca-client enroll" and
the "fabric-ca-client reenroll" commands.  This provides
a greater degree of integration support for fabric by
storing the needed artifacts in MSP format, where
MSP stands for Membership Service Provider.

Change-Id: Ic3c6423fe2df61573847745de6017a7d20b8b16c
Signed-off-by: Keith Smith <bksmith@us.ibm.com>