FAB-2571: FVT for split client/server executables
Updated the existing fvt tests for the new cli commands.

So that the fvt can start running in the CI, tests that are
expected to fail have been placed in staging directory.
This includes tests that
 - are currently failing with known issues
 - need additional code to function properly
Hopefully, this will make it easier to en/disable tests
in the future.

Deleted some convenience stand-alone script files and put
the functionality into the fabric-ca_utils file.

Deleted tests that have been obsoleted by the new configuration.

Updated the Makefile to re-enable the docker-fvt images.

Updated the run_ldap_tests to account for the move of
the ldap directory to ./lib

Updated the start.sh script to set the mysql mode to
'STRICT_TRANS_TABLES'. Otherwise, zero dates are not
permitted. This should require doc updates.

Change-Id: I930a5225e9623fe455217c1d5c2fd0d627617bae
Signed-off-by: Allen Bailey <eabailey@us.ibm.com>