[FAB-2896] Support multiple CAs - new CA struct
This is the first part in a series of change-sets to add support
for multiple CAs. This change-set is creating the distinction
between server and CA. A new CA struct was created and all the
relevant code from server struct was moved over to CA struct.

There is still a one to one relationship between server and CA,
the next change set will add support for starting multiple instances
of CA in a single server by specifying the 'cafiles' option.

Change-Id: Ie0cc1824f501457147609aa9f8aa591477d104d3
Signed-off-by: Saad Karim <skarim@us.ibm.com>