[FAB-2993] Handle string slices in config
This change set allows a string slice to be specified via
command line flags or environment variables.  For example,
[FAB-2896] needs to specify
a list of file names with the following option:
   --cafiles <CA-config-files>

This also gets rid of some kludges that were previously done
because string slices weren't handled.

Note that there is a bug in viper which I had to work around
in how it handles string slices.  See comments in code.

Change-Id: I23aeb9d39ece403b85a03304f895fc8b31998805
Signed-off-by: Keith Smith <bksmith@us.ibm.com>
Signed-off-by: Saad Karim <skarim@us.ibm.com>