[FAB-2571] Move failing tests to staging directory
So that the fvt can start running in the CI, tests that are
expected to fail have been placed in staging directory.
This includes tests that
   - are currently failing with known issues
   - need additional code to function properly
Hopefully, this will make it easier to en/disable tests in the future.

Also need to update the maxpath for the run_fvt script, since
there is now an additional directory in the tree

Change-Id: I351d9eaf0b767f8f4f1db56a909df2df0c23dd83
Signed-off-by: rennman <eabailey@us.ibm.com>