[FAB-1854] Add files with colons to gitignore
See [FAB-1854].
This change set simply adds the path to the files with colons to
the .gitignore so that git clone will work on windows.  The files
which are ignored were not being used since they are part of the
cabundle part of cfssl which is not being used by fabric-ca-server.

This still needs to be tested on Windows but I have to push it first
before I can clone it to verify.

Change-Id: I9b309135da5ffa68b58041e808bcbbe454468a06
Signed-off-by: Keith Smith <bksmith@us.ibm.com>