[ FAB-2909 ] Fix failing UT for file name too long
In environments in which a container or vm shares a
file system with the host, some OS-specific error
messages do not get proagated back to the guest.
This particular test was failing due to an unexpected
error message, 'protocol error', rather than
'file name too long'.

The solution is to modifiy the test to use an absolute
path that is unique to the guest test image

Change-Id: Ief84755a98bbb1d6f7fb1cd3ccbab7f844179aec
Signed-off-by: Allen Bailey <eabailey@us.ibm.com>