[ FAB-3554 ] Add make target for local CI tests
As a convenience to development, and as a sanity-check for pushing code,
it would be desirable to have a make target that builds and runs all ofi
the associated continuous integration tests on a local branch.

make ci-tests

would run all unit and fvt tests directly in the local development environment.

Change-Id: I81191b534777acf5b59a694439ea28f71b72e613
Signed-off-by: Allen Bailey <eabailey@us.ibm.com>