[FAB-3574] Perform deep copy of config file
See [FAB-3574]

In order to create a non-default CA configuration object,
the missing values in the non-default CA are copied from
the default CA object.  The CopyMissingValues function in
util/struct.go performs this operation.  However, it was doing a
shallow copy, but a deep copy is required.  This change set
changes this to perform a deep copy.

The place where problems showed up were when copying BCCSP
configuration which has a pointer to a single default
configuration which could have a path to the key store.
This path should different for each CA, but because of the
shallow copy, they were using the same keystore path.

Change-Id: I514616b7f40766ae189a6ff5d6a225732e96b4f2
Signed-off-by: Keith Smith <bksmith@us.ibm.com>