[FAB-3574] Fix missing CA config values logic
See [FAB-3574]

Missing values in configuration of additional CA must be
replaced with values from the default CA configuration. If
relative paths are specified in the default CA configuration,
the same relative paths must copied over to additional CAs
if missing before the paths are made absolute.

If absolute paths are specified in the default CA, then the entire
absolute path should be copied over to additional CAs if missing.

Change-Id: Idd82772116448baf33c2efbbd8c3f60feec76b29
Signed-off-by: Saad Karim <skarim@us.ibm.com>