[FAB-3574] Finish fix of multi CA config
See [FAB-3574]

When multiple CAs are configured and relative paths for
fields such as the keystore were specified in the default CA
config, the file paths were converted to absolute paths
before being copied into the CA config.  This is incorrect
and caused all CAs to use the same file paths.

The fix for this required breaking the initialization logic
into the following pieces:
1) Initialize the config, setting default values, etc.
   This is done by the server.initConfig function.
2) At the end of this function, perform all initialization
   of the multiple CA configs, copying the relative paths,
   and converting the CA configs paths to absolute.
3) Complete initialization of the default CA config, including
   initially converting the default paths to absolute.

Change-Id: I2ada0c2b1bd31609799d09d422137808188a61d5
Signed-off-by: Keith Smith <bksmith@us.ibm.com>