[FAB-3100] Fix panic in server start
See [FAB-3100] for details on the panic.
The issue is that the net.Listen call periodically returns
nil, nil; in other words, it does not return an error, nor
does it return a listener.  The fix is to retry up to 10
times when this is detected before failing.
This change set also removes all of the sleeps in the tests
so that we are not masking the problem.

Change-Id: I0142d3481bcd6f9685071edf4c5e6b87b4ec1333
Signed-off-by: Keith Smith <bksmith@us.ibm.com>