[ FAB-1383 ] Add TLS tests for mysql
This completes list for TLS enabled tests for currently supported backend DBs

mysql does not support ecdsa certs, so I had to punt. The root CA used in
testdata does not have an RSA signing key. The server provides
a tool that exploits openssl to build it's own hierarchy, so I added the
relevant command to the Dockerfile. I then added the generated rootcert to the
DB.tls.certfiles list.  This restricts testing to the fvt image, but it's
likely to only be run there anyway.

During testing, I also noticed a PATH issue with check_lint, so I fixed it.

Change-Id: Ia3953182cc68b096e61585831f57350f4725c787
Signed-off-by: rennman <eabailey@us.ibm.com>