[FAB-3883] Intermediate CA restriction on CN
When starting up a CA, in the CSR section a CN can
be specified that will be used to generate a
self-signed certificate if a root CA. However, when
a CA is acting like an intermediate it should
not be able to specify CN.

An intermediate CA, acting like a client, will enroll
with the root CA by providing its enrollment ID and
password. The enrollment ID will take place of the CN
in the certificate. An error will be thrown if an
intermediate CA tries to specify a CN. 

See [FAB-3883] for more info

Change-Id: I993d8837a09a4dbbdb625c7cb901798a871ef97d
Signed-off-by: Saad Karim <skarim@us.ibm.com>