[FAB-3121] Improve Intermediate CA error message
When initializing an intermediate CA, it requires
enrolling with a root CA to get the intermediate
CA's signing cert.  If the "-u" option is used
to specify the URL of the root CA, a username
and password must be part of the URL.  If it is
missing, return a better error message.

I also updated the doc to make this clearer.

Change-Id: Ic31a8f6adbd3cacf298bba4f3d85eda13ec0e532
Signed-off-by: Keith Smith <bksmith@us.ibm.com>