[ FAB-3982 ] TLS certificate for fabric-ca tests
There is really less code here than meets the eye.
Primarily it is a matter of populating the fabric-ca-fvt
docker image with some tools that already exist in the
testdata and/or fvt/util directory.

What is new, is the tls_pki.sh script will exploit the
existing pki tool to create an RSA certificate hierarchy for TLS:

  rootCA (20-yr) FabricTlsRootCa 4096-bit
    |
    intermediateCA (10-yr) FabricTlsSubCa 4096-bit
      |
      resourceAuthorityCA (5-yr) FabricTlsRa 4096-bit
        |                      |
        server (2-yr) 2048-bit client (2-yr) 2048-bit
        FabricTlsServerEE      FabricTlsClientEE

RSA was selected as the lowest common denominator
since mysql does not support EC keys.

The auth chain will be bundled into a single file
for easier configuration:   FabricTlsPkiBundle.pem

All of the existing references to certfiles/keyfile in the TLS
configuration (and command line options) have been updated to use
those which are now dynamically generated when the docker image is built.

The crypto artifacts are created in /etc/hyperledger/fabric-ca/,
along with the ldif files and setup scripts. The certificate and key
files will be copied to each server's respctive directory (or in the
case of LDAP, added to the config's DIT)

Lastly, the Dockerfile has been optimized, as the server setup steps have
been moved to individual files. This should speep up the builds and
reduce the size of the image.

These changes have been tested on both s390x and x86_64

Change-Id: I1e25876fa0a4e1370dcff4b301b82eb4cc2754d9
Signed-off-by: rennman <eabailey@us.ibm.com>