[FAB-4096] Remove openldap tarball
This change set deletes the openldap tar file and separate docker
file.  We no longer need the separate openldap image anyway since LDAP,
MySQL, and Postgres are all built dynamically into a single image
as part of fvt testing.  This is part of CI, so it is assured to
run on all platforms.

Change-Id: Ib471b88955c37460a7b2f1bc88f184efc3fce401
Signed-off-by: Keith Smith <bksmith@us.ibm.com>