[ FAB-3924 ] - Additional test coverage for tcert
Brings coverage for tcerts to 90+ %

There are some tests that have been commented
out due to issues that would cause them to fail.

They can be enabled later.

I also modified GenNumber() to check for error, since
I was able to create a panic in that function.
I can't write a unit test for it, since it requires
root authority (or elevated permissions to /dev/urandom)
which is not possible in this environment, so I just fixed it.

Change-Id: Ia452036b367bd509993deda96622d89c9296f0bf
Signed-off-by: Allen Bailey <eabailey@us.ibm.com>