[ FAB-3982 ] TLS dynamic certs for fabric-ca tests
Adds the tls_pki.sh script will exploit the existing pki tool to create an
RSA certificate hierarchy for TLS:
   rootCA (20-yr) FabricTlsRootCa 4096-bit
     |
     intermediateCA (10-yr) FabricTlsSubCa 4096-bit
       |
       resourceAuthorityCA (5-yr) FabricTlsRa 4096-bit
          |                      |
          server (2-yr) 2048-bit client (2-yr) 2048-bit
          FabricTlsServerEE      FabricTlsClientEE

RSA was selected as the lowest common denominator since mysql does not
support EC keys.

The auth chain will be bundled into a single file for
easier configuration: FabricTlsPkiBundle.pem

All of the existing references to certfiles/keyfile in the TLS
configuration (and command line options) have been updated to use those
which are now dynamically generated when the docker image is built.

The crypto artifacts are created in /etc/hyperledger/fabric-ca/,
along with the ldif files and setup scripts.  The certificate and
key files will be copied to each server's respctive directory
(or in the case of LDAP, added to the config's DIT).

Lastly, the Dockerfile has been optimized, as the server
setup steps have been moved to individual files. This should speep up
the builds and reduce the size of the image.

Change-Id: I192508cc34c4f8245d58741b6b0ff1d42dbe7109
Signed-off-by: Allen Bailey <eabailey@us.ibm.com>