[FAB-1989] Fix leaking authority to delegates
The roles and delegates roles were not being checked correctly during
registration, thus allowing an identity to gain privileges they should
not have.  For example, a registrar's delegate roles were incorrectly
allowed to be a superset of the registrar's roles.  This allowed a
user1 to register user2 who could in turn create user3 with more privileges
than user1.  This change set fixes this problems and adds quite a few
test cases for various registration scenarios.

Change-Id: Ic2284d48a9efb2a0553bcacffe386d341cb2fa70
Signed-off-by: Keith Smith <bksmith@us.ibm.com>