[FAB-4093] Fix the TLS client using BCCSP
The TLS client fails when client authentication is enabled and
the key was generated by BCCSP.  This is because the key file
name can not be found since its name is the hash of the public
key.  This change set fixes this for the TLS client in the same
way that https://gerrit.hyperledger.org/r/#/c/9235 fixed this
for the TLS server.  It reuses the util.LoadX509KeyPair function
for the client as is already used for the server.

See TestTLSClientAuth in lib/client_whitebox_test.go for the test
case.

Change-Id: I61432042e5fde3e300e855cd60499862a90cd715
Signed-off-by: Keith Smith <bksmith@us.ibm.com>