[FAB-3228] Fix/clarify CA pathlen constraints
The name of the pathlen fields were wrong in the default config
file.  This change set corrects the field names and adds documentation
to the generated config file and the user's guide to describe the values.
It also adds a test case which shows that by default a root CA allows
intermediate CAs to be enrolled, but enrolling an intermediate CA
with an intermediate CA will fail (by default).

The pathlen constraints of a CA allow the depth of a CA hierarchy
to be limited.  The CA hierarchy refers to the root CA having
child intermediate CAs which can in turn have child intermediate CAs.

Change-Id: I1ed202946a738ae49e91a94cca312c0b393d2325
Signed-off-by: Keith Smith <bksmith@us.ibm.com>