[FAB-3228] Fix/clarify CA pathlen constraints
The name of the pathlen fields were wrong in the default config
file.  This change set corrects the field names and adds documentation
to the generated config file to describe the values.

The pathlen constraints of a CA allow the depth of a CA hierarchy
to be limited.  The CA hierarchy refers to the root CA having
child intermediate CAs which can in turn have child intermediate CAs.

Change-Id: I1ed202946a738ae49e91a94cca312c0b393d2325
Signed-off-by: Keith Smith <bksmith@us.ibm.com>