[FAB-4372] Value of attributes are ignored
There are areas where decisions are made based on
attributes of a registered identity. However, the
current logic only checks to see if the identity
has a specific attribute. There was no check being
made as to the actual value of the attribute. This
was allowing certain execution of code that should
have been prevented if the attribute had a value of
false.

Code was added to make sure that the value of an
attribute is examined before proceeding further.

For reproduction steps and more info, see [FAB-4372]

Change-Id: I0f0b0461d7f4833211afd724e2e1a18b822372ee
Signed-off-by: Saad Karim <skarim@us.ibm.com>