[FAB-4515] Fix concurrency issue with enroll
There was an issue with the logic being used
to update the state of a user (which
reflects the number of times the user
has enrolled).  In the case of MySQL, this
resulted in the error in the [FAB-4515] but
also allowed for a user to exceed max
enrollments when using sqlite3 as the
database.

The fix is pretty straightforward - it moves
the logic into the SQL statement in order to
avoid / depend on locks and or the concurrency
mode set for a database.

Change-Id: Ifeee607049b8104d2b5eb30bb00a8158ff149671
Signed-off-by: Gari Singh <gari.r.singh@gmail.com>