[FAB-4462] Scripts to dynamically generate readme
Added script that will pull in CLI flags from client
and server and create a RST file, the same will also be done
for the configuration file templates. The main fabric-ca
readme will then point to these files as appropriate.

This will remove the need to manually update the doc
everytime a new configuration options is added to
fabric-ca.

Change-Id: I06f5d78e59305b9f0e3272068c5fbf93bac53483
Signed-off-by: Saad Karim <skarim@us.ibm.com>