[FAB-4567] Fix for id attributes security issue
This change set makes a server-side check to require that the
CN (Common Name) in the CSR (Certificate Signing Request) on
an enroll request be the same as the enrollment ID.
The check was previously only made on the client-side, leaving
a security hole which could be exploited by a roque client.

The testMasqueradeEnroll test fails prior to this change and
passes with this change.

Change-Id: Id7c87e6958f5df19ca6308dd9ef8009a5f4a1d74
Signed-off-by: Keith Smith <bksmith@us.ibm.com>