[ FAB-4606 ] Bad intermediate CA CSR
When acting as an intermediate CA, fabric ca server is allocating
an empty CSR and essentially setting it to not nil if it is nil.

Then there is later check in clientconfig.go which checks to see
if Enrollment.CSR is nil; however it will not be nil due to the
previous allocation of Enrollment.CSR, thus the assignment to the
client config CSR never happens.

So an empty CSR gets passed around rather than the correct CSR
that comes from client config.

Consequently, the root server does not recognize this as an
intermediate ca request. This has repercussions for later checks
in the code (e.g. is the hf.IntermediateCA flag set)

Fix is to delete these three lines in lib/ca.go

Change-Id: I729dd212bbae09bd08125aef97fad6803ba9af0b
Signed-off-by: rennman <eabailey@us.ibm.com>