[ FAB-4606 ] Bad intermediate CA CSR
When acting as an intermediate CA, fabric ca server incorrectly nulls out the fields of the CSR, after they are copied over by command initialization:

if clientCfg.Enrollment.CSR == nil {
  clientCfg.Enrollment.CSR = &api.CSRInfo{}
}

Consequently, the root server does not recognize this as an intermediate ca request. This has repercussions for later checks in the code (e.g. is the hf.IntermediateCA flag set)

Fix is to delete these three lines in lib/ca.go

Change-Id: I729dd212bbae09bd08125aef97fad6803ba9af0b
Signed-off-by: rennman <eabailey@us.ibm.com>